Add server side js
Splits the webpack build into a web and node build.

For NodeJS applications this lib can be used as a simple API abstraction layer.

This change adds the first validateToken() method.
