+++ package.json
@@ -25,6 +25,7 @@
     "webpack-cli": "^3.3.6"
   },
   "dependencies": {
+    "axios": "^0.19.0",
     "qrcode": "^1.4.1"
   }
 }
