+++ webpack.config.js
@@ -1,11 +1,11 @@
 const path = require('path');
 
-module.exports = {
+const browserConfig = {
   target: 'web',
   mode: 'production',
-  entry: './src/index.js',
+  entry: './src/browser.js',
   output: {
-    filename: 'index.js',
+    filename: 'browser.js',
     path: path.resolve('dist'),
     library: 'nutsAuth',
   },
@@ -20,5 +20,30 @@ module.exports = {
   },
   resolve: {
     extensions: ['.js'],
-  },
+  }
 };
+
+const nodeConfig = {
+  target: 'node',
+  mode: 'production',
+  entry: './src/index.js',
+  output: {
+    filename: 'index.js',
+    path: path.resolve('dist'),
+    libraryTarget: "commonjs",
+  },
+  module: {
+    rules: [
+      {
+        test: /\.js?$/,
+        exclude: /(node_modules)/,
+        use: 'babel-loader',
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['.js'],
+  }
+};
+
+module.exports = [browserConfig, nodeConfig];
